Fixes null check in recycle bin repo
Conflicts:
	src/Umbraco.Core/Persistence/Repositories/RecycleBinRepository.cs

+++ src/Umbraco.Core/Persistence/Repositories/RecycleBinRepository.cs
@@ -131,18 +131,19 @@ namespace Umbraco.Core.Persistence.Repositories
             {
                 var fs = FileSystemProviderManager.Current.GetFileSystemProvider<MediaFileSystem>();
                 Parallel.ForEach(files, file =>
-                                        {
-                                            if (UmbracoSettings.UploadAllowDirectories)
-                                            {
-                                                var relativeFilePath = fs.GetRelativePath(file);
-                                                var parentDirectory = System.IO.Path.GetDirectoryName(relativeFilePath);
-                                                fs.DeleteDirectory(parentDirectory, true);
-                                            }
-                                            else
-                                            {
-                                                fs.DeleteFile(file, true);
-                                            }
-                                        });
+                {
+                    if (file.IsNullOrWhiteSpace()) return;
+                    if (UmbracoSettings.UploadAllowDirectories)
+                    {
+                        var relativeFilePath = fs.GetRelativePath(file);
+                        var parentDirectory = System.IO.Path.GetDirectoryName(relativeFilePath);
+                        fs.DeleteDirectory(parentDirectory, true);
+                    }
+                    else
+                    {
+                        fs.DeleteFile(file, true);
+                    }
+                });
 
                 return true;
             }
